silence signed vs. unsigned integer warnings

+++ src/MISC/dump_xtc.cpp
@@ -555,7 +555,7 @@ static int sizeofint(const int size)
   unsigned int num = 1;
   int num_of_bits = 0;
 
-  while (size >= num && num_of_bits < 32) {
+  while (size >= (int) num && num_of_bits < 32) {
     num_of_bits++;
     num <<= 1;
   }
@@ -596,7 +596,7 @@ static int sizeofints( const int num_of_ints, unsigned int sizes[])
   }
   num = 1;
   num_of_bytes--;
-  while (bytes[num_of_bytes] >= num) {
+  while ((int)bytes[num_of_bytes] >= num) {
     num_of_bits++;
     num *= 2;
   }
@@ -650,7 +650,7 @@ static void sendints(int buf[], const int num_of_ints, const int num_of_bits,
     }
     num_of_bytes = bytecnt;
   }
-  if (num_of_bits >= num_of_bytes * 8) {
+  if (num_of_bits >= (int)num_of_bytes * 8) {
     for (i = 0; i < num_of_bytes; i++) {
       sendbits(buf, 8, bytes[i]);
     }
